+++ desc/objectives/getters.py
@@ -26,7 +26,7 @@ from .linear_objectives import (
     FixPressure,
     FixPsi,
 )
-from .nae_utils import make_RZ_cons_1st_order
+from .nae_utils import calc_zeroth_order_lambda, make_RZ_cons_1st_order
 from .objective_funs import ObjectiveFunction
 
 
@@ -192,6 +192,7 @@ def get_NAE_constraints(
     kinetic=False,
     normalize=True,
     N=None,
+    fix_lambda=False,
 ):
     """Get the constraints necessary for fixing NAE behavior in an equilibrium problem.
 
@@ -212,9 +213,13 @@ def get_NAE_constraints(
         Whether to also fix kinetic profiles.
     normalize : bool
         Whether to apply constraints in normalized units.
-    N : int,
+    N : int
         max toroidal resolution to constrain.
         If None, defaults to equilibrium's toroidal resolution
+    fix_lambda : bool
+        Whether to constrain lambda to match that of the NAE near-axis
+        if an int, fixes lambda up to that order in rho {0,1}
+        if True, fixes lambda up to the specified order given by order
 
     Returns
     -------
@@ -257,8 +262,15 @@ def get_NAE_constraints(
             constraints += (
                 FixCurrent(eq=desc_eq, normalize=normalize, normalize_target=normalize),
             )
+    if fix_lambda or fix_lambda >= 0:
+        L_axis_constraints, _, _ = calc_zeroth_order_lambda(
+            qsc=qsc_eq, desc_eq=desc_eq, N=N
+        )
+        constraints += L_axis_constraints
     if order >= 1:  # first order constraints
-        constraints += make_RZ_cons_1st_order(qsc=qsc_eq, desc_eq=desc_eq, N=N)
+        constraints += make_RZ_cons_1st_order(
+            qsc=qsc_eq, desc_eq=desc_eq, N=N, fix_lambda=fix_lambda and fix_lambda > 0
+        )
     if order >= 2:  # 2nd order constraints
         raise NotImplementedError("NAE constraints only implemented up to O(rho) ")